+++ xxl-job-admin/src/test/java/com/xxl/job/admin/controller/JobInfoControllerTest.java
@@ -1,52 +1,47 @@
 package com.xxl.job.admin.controller;
 
-import com.xxl.job.admin.core.model.XxlJobInfo;
-import org.codehaus.jackson.map.ObjectMapper;
+import com.xxl.job.admin.controller.interceptor.PermissionInterceptor;
+import com.xxl.job.admin.core.util.PropertiesUtil;
 import org.junit.Before;
 import org.junit.Test;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MvcResult;
+import org.springframework.util.LinkedMultiValueMap;
+import org.springframework.util.MultiValueMap;
 
 import javax.servlet.http.Cookie;
 
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 
 public class JobInfoControllerTest extends AbstractSpringMvcTest {
-  Cookie cookie;
+
+  private Cookie cookie;
 
   @Before
   public void login() throws Exception {
     MvcResult ret = mockMvc.perform(
         post("/login")
             .contentType(MediaType.APPLICATION_FORM_URLENCODED)
-            .param("userName", "admin")
-            .param("password", "123456")
+            .param("userName", PropertiesUtil.getString("xxl.job.login.username"))
+            .param("password", PropertiesUtil.getString("xxl.job.login.password"))
     ).andReturn();
-    cookie = ret.getResponse().getCookie("LOGIN_IDENTITY");
+    cookie = ret.getResponse().getCookie(PermissionInterceptor.LOGIN_IDENTITY_KEY);
   }
 
   @Test
   public void testAdd() throws Exception {
-    XxlJobInfo jobInfo = new XxlJobInfo();
-    jobInfo.setJobGroup(1);
-    jobInfo.setJobDesc("autoEnquiryStatisPerWeek");
-    jobInfo.setExecutorRouteStrategy("FIRST");
-    jobInfo.setJobCron("0 0 1 ? * MON");
-    jobInfo.setGlueType("BEAN");
-    jobInfo.setExecutorHandler("AutoEnquriy");
-    jobInfo.setExecutorBlockStrategy("SERIAL_EXECUTION");
-    jobInfo.setExecutorFailStrategy("FAIL_ALARM");
-    jobInfo.setAuthor("listening");
+    MultiValueMap<String, String> parameters = new LinkedMultiValueMap<String, String>();
+    parameters.add("jobGroup", "1");
 
-    ObjectMapper mapper = new ObjectMapper();
-    String jobInfoStr = mapper.writeValueAsString(jobInfo);
     MvcResult ret = mockMvc.perform(
-        post("/jobinfo/add")
+        post("/jobinfo/pageList")
             .contentType(MediaType.APPLICATION_FORM_URLENCODED)
-            .content(jobInfoStr)
+            //.content(paramsJson)
+            .params(parameters)
             .cookie(cookie)
     ).andReturn();
 
     System.out.println(ret.getResponse().getContentAsString());
   }
+
 }
