+++ xxl-job-admin/src/test/java/com/xxl/job/admin/dao/XxlJobLogDaoTest.java
@@ -1,4 +1,4 @@
-package com.xxl.job.dao.impl;
+package com.xxl.job.admin.dao;
 
 import com.xxl.job.admin.core.model.XxlJobLog;
 import com.xxl.job.admin.dao.XxlJobLogDao;
